Fix scheduler issues
It was reported to me that last Friday Punssi-Buddy didn't do the
announacements. One reason for this might be that eventually the
scheduled jobs are garbage collected or something, once the application
has been running for hours or something.

One solution to this would be storing the scheduled jobs into a
container that isn't garbage collected at any point. For this reason, I
decided to store them all in an array, and according to my initial tests
it seems to work.

Also add a script entry into the `package.json` so that the application
can be started from the repository with `npm start` command.